Refactor cpp files into directories
