+++ app/src/main/cpp/CMakeLists.txt
@@ -1,18 +1,18 @@
 include(cmake/list_transform.cmake)
 set(SYNTH_ENGINE_SOURCES
-        AudioEngine.cpp
-        AudioStream.cpp
-        Envelope.cpp
-        EnvelopeControlledAmplifier.cpp
-        Filter.cpp
-        native-lib.cpp
-        NoiseWaveform.cpp
-        Oscillator.cpp
-        PlayException.cpp
-        SineWaveform.cpp
-        SquareWaveform.cpp
-        TriangleWaveform.cpp
-        Waveform.cpp
+        jni/native-lib.cpp
+        synth/AudioEngine.cpp
+        synth/AudioStream.cpp
+        synth/Envelope.cpp
+        synth/EnvelopeControlledAmplifier.cpp
+        synth/Filter.cpp
+        synth/NoiseWaveform.cpp
+        synth/Oscillator.cpp
+        synth/PlayException.cpp
+        synth/SineWaveform.cpp
+        synth/SquareWaveform.cpp
+        synth/TriangleWaveform.cpp
+        synth/Waveform.cpp
         )
 add_library(
         synth-engine
