+++ app/src/main/cpp/jni/native-lib.cpp
@@ -1,12 +1,12 @@
-#include "AudioEngine.h"
-#include "AudioStream.h"
-#include "Envelope.h"
-#include "Filter.h"
-#include "NoiseWaveform.h"
-#include "Oscillator.h"
-#include "SineWaveform.h"
-#include "SquareWaveform.h"
-#include "TriangleWaveform.h"
+#include "../synth/AudioEngine.h"
+#include "../synth/AudioStream.h"
+#include "../synth/Envelope.h"
+#include "../synth/Filter.h"
+#include "../synth/NoiseWaveform.h"
+#include "../synth/Oscillator.h"
+#include "../synth/SineWaveform.h"
+#include "../synth/SquareWaveform.h"
+#include "../synth/TriangleWaveform.h"
 #include <jni.h>
 #include <oboe/Oboe.h>
 #include <string>
