+++ app/src/main/cpp/synth/Envelope.cpp
