+++ app/src/main/cpp/synth/Envelope.h
