+++ app/src/main/cpp/synth/EnvelopeControlledAmplifier.cpp
