+++ app/src/main/cpp/synth/EnvelopeControlledAmplifier.h
