+++ app/src/main/cpp/synth/EnvelopeParameters.h
