+++ app/src/main/cpp/synth/PlayException.cpp
