+++ app/src/main/cpp/synth/PlayException.h
